Rendering core (Jet-derived): direct triangle rasterizer (DIRECT_RASTERIZER,
legacy QuadTree fallback) + vertex pre-transform + render stats + on-device
RASTER_VERIFY_AB gate + FastTrig LUT. render ~20.8ms -> ~6ms on device.

Face/morph memory: cull morphs unused by the active animation config (M1);
half-float morph deltas + uint16 indices (M3); uint16 IndexGroup (M4); render
scratch + z-buffer moved to PSRAM (frees ~40KB internal DRAM; intFree ~20K ->
~46K, largest block 7.7K -> 31.7K).

Firmware/internal-SRAM analysis reports + SRAM_DUMP diag. Face download uses
freshly-fetched device_id (config-first). device_id now always chip-internal
eFuse unique ID. Fixed validate.ps1 Check-Order fragility + merge-bin esptool
v5 + TriangleGroup latent double-free. host-tests/ A/B harness.

+++ src/Render/IndexGroup.h
@@ -1,10 +1,16 @@
 #pragma once
 
+#include <cstdint>
+
+// Triangle vertex indices. Stored as uint16 (vertex counts are < 65536; the
+// face has 495). This halves the index buffer (6276 B -> 3138 B for 523 tris)
+// and is read only at mesh-construction time (the render hot path uses the
+// already-wired vertex pointers), so narrowing is render-path neutral.
 typedef struct IndexGroup {
 public:
-	unsigned int A = 0;
-	unsigned int B = 0;
-	unsigned int C = 0;
+	uint16_t A = 0;
+	uint16_t B = 0;
+	uint16_t C = 0;
 
     IndexGroup() {
         this->A = 0;
@@ -19,45 +25,45 @@ public:
     }
 
     IndexGroup(unsigned int X, unsigned int Y, unsigned int Z) {
-        this->A = X;
-        this->B = Y;
-        this->C = Z;
+        this->A = (uint16_t)X;
+        this->B = (uint16_t)Y;
+        this->C = (uint16_t)Z;
     }
 
     IndexGroup Add(IndexGroup indexGroup) {
         return IndexGroup {
-            this->A + indexGroup.A,
-            this->B + indexGroup.B,
-            this->C + indexGroup.C 
+            (uint16_t)(this->A + indexGroup.A),
+            (uint16_t)(this->B + indexGroup.B),
+            (uint16_t)(this->C + indexGroup.C)
         };
     }
 
     IndexGroup Subtract(IndexGroup indexGroup) {
         return IndexGroup {
-            this->A - indexGroup.A,
-            this->B - indexGroup.B,
-            this->C - indexGroup.C 
+            (uint16_t)(this->A - indexGroup.A),
+            (uint16_t)(this->B - indexGroup.B),
+            (uint16_t)(this->C - indexGroup.C)
         };
     }
 
     IndexGroup Multiply(IndexGroup indexGroup) {
         return IndexGroup {
-            this->A * indexGroup.A,
-            this->B * indexGroup.B,
-            this->C * indexGroup.C 
+            (uint16_t)(this->A * indexGroup.A),
+            (uint16_t)(this->B * indexGroup.B),
+            (uint16_t)(this->C * indexGroup.C)
         };
     }
 
     IndexGroup Divide(IndexGroup indexGroup) {
         return IndexGroup {
-            this->A / indexGroup.A,
-            this->B / indexGroup.B,
-            this->C / indexGroup.C 
+            (uint16_t)(this->A / indexGroup.A),
+            (uint16_t)(this->B / indexGroup.B),
+            (uint16_t)(this->C / indexGroup.C)
         };
     }
-    
+
     String ToString() {
         return "[" + String(A) + ", " + String(B) + ", " + String(C) + "]";
     }
-    
+
 } IndexGroup;
